refactor: remove unused code and interfaces
This commit removes unused code and interfaces that were not being used in the project. The CustomerDao interface was replaced by the CustomerDaoCrudRepository interface, which is a Spring Data interface that provides basic CRUD operations. The CustomerDaoImpl class was also removed since it was not being used. Finally, the ICrudDao interface was removed since it was not being used in the project.